Heidelberger and Welch
For large n, Bn converges in distribution to a Brownian bridge (Billingsley 1986). So you can construct a test statistic by using Bn. The statistic used in these procedures is the Cramer–von Mises statistics.  As n → infinite, the statistic converges in distribution to a standard Cramer–von Mises distribution.
And the Cramer-Lundberg approximation is an approximation of the Cramer-von Mises test and is used to calculate the $p$-value of the Cramer-von Mises statistic.