Create 8. KL divergence tool
The main idea of Tool 1 is to compare the distribution of the MCMC chain with the target distribution. If the chain has reached equilibrium, these two distributions should be very similar to each other.

The comparison is done through Kullback-Leibler (KL) divergence, which is a measure of how one probability distribution diverges from a second, expected probability distribution. It is an asymmetric measure of the difference between two probability distributions P and Q.